+++ spec/spec_helper.rb
@@ -31,6 +31,7 @@ require "rspec/its"
 require "rspec/collection_matchers"
 require "delorean"
 require "tempfile"
+require "pry"
 
 # Requires supporting ruby files with custom matchers and macros, etc,
 # in spec/support/ and its subdirectories.